+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to the ld-find-code-refs program will be documented in this 
 
 ## [1.3.0] - [DATE]
 
+### Added
+
+- Added a `--outDir` option to the CLI. If provided, code references will be written to a csv file in `outDir`.
+- Added a `--dryRun` option to the CLI. If provided, `ld-find-code-refs` will scan for code references without sending them to LaunchDarkly. May be used in conjunction with `--outDir` to output code references data to a csv file instead of sending data to LaunchDarkly.
+
 ### Fixed
+
 - `ld-find-code-refs` now supports scanning repositories with a large number of flags using a pagination strategy. Thanks @cuzzasoft!
 - Delimiters will now always be respected when searching for flags referenced in code. This fixes a bug causing references for certain flag keys to match against other flag keys that are substrings of the matched reference.